pyproject.toml
[project]
name = "recursive-dialectical-engine"
version = "0.1.0"
description = "Multi-model dialectical reasoning with recursive arbitration"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
local = [
    "mlx>=0.22.0",
    "mlx-lm>=0.22.0",
]
cloud = ["anthropic>=0.40.0", "openai>=1.50.0", "google-generativeai>=0.8.0"]
sandbox = ["modal>=0.70.0", "e2b-code-interpreter>=1.0.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.8"]

[project.scripts]
rde = "rde.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rde"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring API keys (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
